Subject: Added welcome message to MainMenu.cs
Body:
A new line of code has been introduced in the `MainMenu.cs` file within the `CarSimulator.Menus` namespace. This line outputs a welcome message in Swedish, "Välkommen till bilkörningssimulatorn!", to the console when the main menu of the application is loaded. This translates to "Welcome to the car driving simulator!" in English.