Updated test classes accessibility and attributes
Modified the accessibility of several classes in the `CarSimulatorTests` and `LibraryTests.Services` namespaces from `internal` to `public` to allow access outside their current assembly. Also, added the `[TestClass]` attribute to these classes to specify that they contain methods for the test engine to execute. No changes were made to the methods or properties within these classes.