Refactor FoodService to use IConsoleService for testability
Refactored the `FoodService` class to use an instance of `IConsoleService` instead of directly using the `Console` class, improving testability by allowing console operations to be mocked in unit tests. The `FoodService` constructor now requires an additional `IConsoleService` parameter, which is used for all console operations in the class. Updated the `CreateActionService` method in the `ActionServiceFactory` class to pass an `IConsoleService` instance to the `FoodService` constructor. Replaced all direct calls to `Console` methods in the `FoodService` class with calls to corresponding `_consoleService` methods. Added `using` directives for `System` and `System.Collections.Generic` to `FoodService`. The `FoodService` class now throws an `ArgumentNullException` if the provided `IConsoleService` instance is `null`, preventing potential null reference exceptions.