Refactor MainMenuService to use IConsoleService
Refactored `MainMenuService` to use an instance of `IConsoleService` for improved testability and separation of concerns. Replaced all direct calls to `Console` methods in `MainMenuService` with calls to equivalent methods on the `_consoleService` instance. Updated the `MainMenuService` constructor to throw an `ArgumentNullException` if the provided `IConsoleService` instance is null. Also, modified the `AppRun` method in `AppStart.cs` to create an instance of `MainMenuService` with an `IConsoleService` argument.