Refactor MenuDisplayService to use IConsoleService
Refactored the `MenuDisplayService` class to use an instance of `IConsoleService` instead of the `Console` class for better unit testing and separation of concerns. Updated `AppStart` to pass an `IConsoleService` instance to the `MenuDisplayService` constructor. Replaced all direct calls to `Console` methods in `MenuDisplayService` with calls to equivalent `IConsoleService` methods. Updated various methods in `MenuDisplayService` to use `IConsoleService` methods instead of equivalent `Console` methods.