Updated MainMenu class to use IConsoleService
The MainMenu class and its instantiation in AppStart.cs have been updated to include a dependency on IConsoleService, replacing all direct calls to the Console class with calls to the equivalent methods in _consoleService. This change enhances unit testing and separation of concerns. A null check for the consoleService parameter has been added in the MainMenu constructor to ensure a valid IConsoleService instance is always provided. The changes mainly relate to dependency injection and the use of an interface to abstract console operations, making the code more testable and modular.