`Added 'Title' property to 'Driver' class and updated related services and tests`
This commit includes the addition of a new 'Title' property to the 'Driver' class and updates to the 'Name' property to include the 'Title'. The 'MainMenuService' and 'RandomUserService' have been updated to include the 'Title' when displaying the driver's name and when creating a new 'Driver' object.

The 'MainMenuTests' and 'MainMenuServiceTests' have been updated to include the 'Title' in the 'driver' object and to expect the 'Title' to be included when the driver's name is displayed. The 'Title' is now also checked in the assertions.