`Add Title property test in RandomUserServiceTests.cs`
` `

`Added a new Title property to the Name object in the Result object in RandomUserServiceTests.cs for testing purposes. Also, a new assertion has been added to the test method to check if the Title property of the result object equals "Mr". This is to ensure the Title property is correctly set and returned by the GetRandomDriverAsync method.`